Add VTT TSI tables in TTX format in UFO3's data folder; store VTT-computed maxp values in UFO lib.plist
We use 'com.robofont.robohint.maxp.*' lib keys to store maxp values and these are used by Robofont as well.

The VTT sources were dumped from Ubuntu 0.80 VTT source TTFs, using vttdump script.
This works the same as ttx, but it also strips the timestamps and the glyph indexes that appear inside VTT
comments. These make the git diff change every time one compiles the fonts with VTT, or when the glyph order
changes in the font, but not the actual hinting istructions' content.