BI, LI, MI and M: set hdmx table values for uniF000 to zero
uniF000 is the debug glyph looking like a LCD-style 88 for testing if
and what hinting is used. In the Google Fonts release of Ubuntu-BI, LI,
MI and M, all fields were set to zero while CacheTT said something else.
To establish a baseline, do as the GF release does.

REMOVE THIS HACK ONCE IT HAS BEEN PROBEN UNNECESSARY.