Match B, BI, C, L, LI, M, MI GPOS table to GF releases
Turns out the GPOS and kern tables disagree on some Greek kernings and
quite a lot of kernings in C. This commit restores GPOS kerning equality
between the UFO sources and the GF releases. The C kerning that was
previously stripped is rolled back.

The kern table will simply be fixed up by post-processing.