Ubuntu-M: Fix displaced caron on uni01EE
The caron was being shifted far to the right in the code that ran after
both components had been hinted. Compared to the GF release, the
composite had one point less without the code being adjusted. The second
CALL (to function 80 which centers accents on their base glyph) then
centered the wrong point and SHC[rp0] subsequently moved the rest of the
caron too far to the right.